FromDigits

Documentation for FromDigits.jl.

Examples

julia> using FromDigits

julia> digits(739)
3-element Vector{Int64}:
 9
 3
 7

julia> fromdigits([9, 3, 7])
739

julia> fromdigits(fill(1, 19))
1111111111111111111

julia> fromdigits(fill(1, 20))
ERROR: OverflowError: 10 * 1111111111111111111 overflowed for type Int64
Stacktrace:
[...]

julia> fromdigits_unchecked(fill(1, 20))
-7335632962598440505

julia> fromdigits(BigInt, fill(1, 20))
11111111111111111111

Library

FromDigits.fromdigits

FromDigits.fromdigits

FromDigits.fromdigits

FromDigits.fromdigits_unchecked

Public

FromDigits.fromdigits — Function

Converts from digits to an integer. This is the inverse function of digits.

Returns the integer represented by the digits in the base. The result is of type T, which defaults to D.

See also fromdigits_unchecked.

FromDigits.fromdigits — Method

fromdigits(digits::AbstractVector{D}; base::D=D(10)) where D<:Integer

Returns the integer represented by the digits in the base. The result is of type D.

FromDigits.fromdigits — Method

fromdigits(T::Type{<:Integer}, digits::AbstractVector{D}; base::D=D(10)) where D<:Integer

Returns the integer represented by the digits in the base. The result is of type T.

FromDigits.fromdigits_unchecked — Function

Converts from digits to an integer. This is the inverse function of digits.

This function does not check the validity of the digits nor overflows in the computations.

See also fromdigits.

Private